// src/api/generated/models/yandexDiskFolder.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * TG Bot Mini App API
 * API for Telegram Mini App settings, OAuth, and usage monitoring
 * OpenAPI spec version: 1.0.0
 */

/**
 * Represents a folder or file in Yandex.Disk.
 */
export interface YandexDiskFolder {
  name: string;
  path: string;
  type: string;
  created?: string | null;
  modified?: string | null;
}